refactor(ScoreConverter): 패키지 이동
